+++ app/features/tournament-bracket/core/tests/mocks.ts
@@ -2097,7 +2097,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709743534,
           members: [
             {
@@ -2208,7 +2208,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709737918,
           members: [
             {
@@ -2332,7 +2332,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709743523,
           members: [
             {
@@ -2443,7 +2443,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709743262,
           members: [
             {
@@ -2554,7 +2554,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709741396,
           members: [
             {
@@ -2678,7 +2678,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709711811,
           members: [
             {
@@ -2802,7 +2802,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709738831,
           members: [
             {
@@ -2926,7 +2926,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709737837,
           members: [
             {
@@ -3037,7 +3037,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709741719,
           members: [
             {
@@ -3174,7 +3174,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709730354,
           members: [
             {
@@ -3298,7 +3298,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709745630,
           members: [
             {
@@ -3419,7 +3419,7 @@ export const PADDLING_POOL_257 = () =>
           prefersNotToHost: 1,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709592381,
           members: [
             {
@@ -3556,7 +3556,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709723749,
           members: [
             {
@@ -3680,7 +3680,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709668399,
           members: [
             {
@@ -3817,7 +3817,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709735267,
           members: [
             {
@@ -3932,7 +3932,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709745849,
           members: [
             {
@@ -4047,7 +4047,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709742258,
           members: [
             {
@@ -4168,7 +4168,7 @@ export const PADDLING_POOL_257 = () =>
           prefersNotToHost: 1,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709738744,
           members: [
             {
@@ -4292,7 +4292,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709746054,
           members: [
             {
@@ -4404,7 +4404,7 @@ export const PADDLING_POOL_257 = () =>
           prefersNotToHost: 1,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709744894,
           members: [
             {
@@ -4515,7 +4515,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709728278,
           members: [
             {
@@ -4626,7 +4626,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709715006,
           members: [
             {
@@ -4737,7 +4737,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709660578,
           members: [
             {
@@ -4852,7 +4852,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709721869,
           members: [
             {
@@ -4977,7 +4977,7 @@ export const PADDLING_POOL_257 = () =>
           prefersNotToHost: 1,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709743633,
           members: [
             {
@@ -5092,7 +5092,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709738747,
           members: [
             {
@@ -5220,7 +5220,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709626047,
           members: [
             {
@@ -5344,7 +5344,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709727951,
           members: [
             {
@@ -5455,7 +5455,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709741482,
           members: [
             {
@@ -5589,7 +5589,7 @@ export const PADDLING_POOL_257 = () =>
           prefersNotToHost: 1,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709744451,
           members: [
             {
@@ -5704,7 +5704,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709726536,
           members: [
             {
@@ -5815,7 +5815,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709558706,
           members: [
             {
@@ -5930,7 +5930,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709744323,
           members: [
             {
@@ -6058,7 +6058,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709677397,
           members: [
             {
@@ -6169,7 +6169,7 @@ export const PADDLING_POOL_257 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1709618711,
           members: [
             {
@@ -7997,7 +7997,7 @@ export const PADDLING_POOL_255 = () =>
           prefersNotToHost: 1,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708476597,
           members: [
             {
@@ -8108,7 +8108,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708535137,
           members: [
             {
@@ -8219,7 +8219,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708533764,
           members: [
             {
@@ -8340,7 +8340,7 @@ export const PADDLING_POOL_255 = () =>
           prefersNotToHost: 1,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708537512,
           members: [
             {
@@ -8464,7 +8464,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708533309,
           members: [
             {
@@ -8575,7 +8575,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708430641,
           members: [
             {
@@ -8699,7 +8699,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708536306,
           members: [
             {
@@ -8807,7 +8807,7 @@ export const PADDLING_POOL_255 = () =>
           prefersNotToHost: 1,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708526368,
           members: [
             {
@@ -8918,7 +8918,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708506060,
           members: [
             {
@@ -9055,7 +9055,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708526814,
           members: [
             {
@@ -9163,7 +9163,7 @@ export const PADDLING_POOL_255 = () =>
           prefersNotToHost: 1,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708466421,
           members: [
             {
@@ -9287,7 +9287,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708377426,
           members: [
             {
@@ -9408,7 +9408,7 @@ export const PADDLING_POOL_255 = () =>
           prefersNotToHost: 1,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708448289,
           members: [
             {
@@ -9545,7 +9545,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708532602,
           members: [
             {
@@ -9656,7 +9656,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708535205,
           members: [
             {
@@ -9780,7 +9780,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708515945,
           members: [
             {
@@ -9891,7 +9891,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708453334,
           members: [
             {
@@ -9999,7 +9999,7 @@ export const PADDLING_POOL_255 = () =>
           prefersNotToHost: 1,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708522730,
           members: [
             {
@@ -10123,7 +10123,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708375443,
           members: [
             {
@@ -10247,7 +10247,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708532665,
           members: [
             {
@@ -10362,7 +10362,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708364254,
           members: [
             {
@@ -10490,7 +10490,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708464101,
           members: [
             {
@@ -10615,7 +10615,7 @@ export const PADDLING_POOL_255 = () =>
           prefersNotToHost: 1,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708520249,
           members: [
             {
@@ -10739,7 +10739,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708535804,
           members: [
             {
@@ -10850,7 +10850,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708535891,
           members: [
             {
@@ -10974,7 +10974,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708521749,
           members: [
             {
@@ -11098,7 +11098,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708536584,
           members: [
             {
@@ -11209,7 +11209,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708537772,
           members: [
             {
@@ -11350,7 +11350,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708379916,
           members: [
             {
@@ -11474,7 +11474,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708519753,
           members: [
             {
@@ -11602,7 +11602,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708534312,
           members: [
             {
@@ -11727,7 +11727,7 @@ export const PADDLING_POOL_255 = () =>
           prefersNotToHost: 1,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708531929,
           members: [
             {
@@ -11838,7 +11838,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708477155,
           members: [
             {
@@ -11953,7 +11953,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708531564,
           members: [
             {
@@ -12107,7 +12107,7 @@ export const PADDLING_POOL_255 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1708503356,
           members: [
             {
@@ -14212,7 +14212,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707443313,
           members: [
             {
@@ -14311,7 +14311,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707366405,
           members: [
             {
@@ -14410,7 +14410,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1706912643,
           members: [
             {
@@ -14509,7 +14509,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707359335,
           members: [
             {
@@ -14634,7 +14634,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707171426,
           members: [
             {
@@ -14746,7 +14746,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707342696,
           members: [
             {
@@ -14858,7 +14858,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707513942,
           members: [
             {
@@ -14983,7 +14983,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707526815,
           members: [
             {
@@ -15108,7 +15108,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707583385,
           members: [
             {
@@ -15207,7 +15207,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707486395,
           members: [
             {
@@ -15306,7 +15306,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707513290,
           members: [
             {
@@ -15405,7 +15405,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707531084,
           members: [
             {
@@ -15504,7 +15504,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707568466,
           members: [
             {
@@ -15616,7 +15616,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707481625,
           members: [
             {
@@ -15715,7 +15715,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707530166,
           members: [
             {
@@ -15814,7 +15814,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707181792,
           members: [
             {
@@ -15926,7 +15926,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707550321,
           members: [
             {
@@ -16042,7 +16042,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707575096,
           members: [
             {
@@ -16154,7 +16154,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707569490,
           members: [
             {
@@ -16279,7 +16279,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707537425,
           members: [
             {
@@ -16378,7 +16378,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707564691,
           members: [
             {
@@ -16507,7 +16507,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707145818,
           members: [
             {
@@ -16614,7 +16614,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707558330,
           members: [
             {
@@ -16713,7 +16713,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707586842,
           members: [
             {
@@ -16812,7 +16812,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707583597,
           members: [
             {
@@ -16937,7 +16937,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707429804,
           members: [
             {
@@ -17049,7 +17049,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707539973,
           members: [
             {
@@ -17165,7 +17165,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707507831,
           members: [
             {
@@ -17277,7 +17277,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707586297,
           members: [
             {
@@ -17386,7 +17386,7 @@ export const IN_THE_ZONE_32 = () =>
           prefersNotToHost: 1,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707583885,
           members: [
             {
@@ -17511,7 +17511,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707578076,
           members: [
             {
@@ -17624,7 +17624,7 @@ export const IN_THE_ZONE_32 = () =>
           prefersNotToHost: 1,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707582953,
           members: [
             {
@@ -17723,7 +17723,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707575330,
           members: [
             {
@@ -17835,7 +17835,7 @@ export const IN_THE_ZONE_32 = () =>
           team: null,
           inviteCode: null,
           activeRosterUserIds: [],
-          avatarUrl: null,
+          pickupAvatarUrl: null,
           createdAt: 1707527645,
           members: [
             {
