+++ app/features/tournament-bracket/core/tests/test-utils.ts
@@ -17,7 +17,7 @@ const tournamentCtxTeam = (
     mapPool: [],
     members: [],
     activeRosterUserIds: [],
-    avatarUrl: null,
+    pickupAvatarUrl: null,
     name: "Team " + teamId,
     prefersNotToHost: 0,
     droppedOut: 0,
